[Refactoring]:  Location 구조체의 위치 변경 & 내가 작성한 메모 불러오기 - MypageViewModel에 있던 Location을 Memo.swift로 이동하였습니다. - 현재 사용자가 작성한 메모를 불러오는 로직을 추가하였습니다. - 마이페이지로 이동할 경우 메모들이 최신화됩니다.
